# [wasm] Speedup Vector128.Shuffle with SIMD