Example speedup result in `Span.Reverse` with `WasmSIMD=true`

Firefox/amd64

| measurement | before | after | delta |
|-:|-:|-:|-:|
|                    Span, Reverse bytes |     0.0857ms |     0.0022ms | 39x faster  |
|                    Span, Reverse chars |     0.0987ms |     0.0042ms | 23.5x faster|

Chrome/amd64

| measurement | before | after | delta |
|-:|-:|-:|-:|
|                    Span, Reverse bytes |     0.0768ms |     0.0028ms | 27.4x faster |
|                    Span, Reverse chars |     0.0975ms |     0.0061ms | 16x faster   |