@lewing 2022-08-31
@fanyang-mono shuffle looks like it would be useful to implement for arm64 as well